requirements.txt
python-dotenv
ollama
sqlalchemy
